+++ x/core/01_interchain_security/types/message_id_multisig.go
@@ -6,7 +6,9 @@ import (
 	"fmt"
 	"slices"
 
+	storetypes "cosmossdk.io/store/types"
 	"github.com/bcp-innovations/hyperlane-cosmos/util"
+	sdk "github.com/cosmos/cosmos-sdk/types"
 )
 
 var _ HyperlaneInterchainSecurityModule = &MessageIdMultisigISM{}
@@ -19,13 +21,16 @@ func (m *MessageIdMultisigISM) ModuleType() uint8 {
 	return INTERCHAIN_SECURITY_MODULE_TYPE_MESSAGE_ID_MULTISIG
 }
 
-func (m *MessageIdMultisigISM) Verify(_ context.Context, rawMetadata []byte, message util.HyperlaneMessage) (bool, error) {
+func (m *MessageIdMultisigISM) Verify(ctx context.Context, rawMetadata []byte, message util.HyperlaneMessage) (bool, error) {
 	metadata, err := NewMessageIdMultisigMetadata(rawMetadata)
 	if err != nil {
 		return false, err
 	}
 
 	digest := metadata.Digest(&message)
+
+	// Charge Cosmos gas for each signature verification.
+	sdk.UnwrapSDKContext(ctx).GasMeter().ConsumeGas(storetypes.Gas(1000*len(metadata.Signatures)), "ism signature verification")
 
 	return VerifyMultisig(m.Validators, m.Threshold, metadata.Signatures, digest)
 }