+++ x/core/01_interchain_security/types/message_id_multisig_test.go
@@ -3,9 +3,9 @@ package types_test
 import (
 	"fmt"
 
+	i "github.com/bcp-innovations/hyperlane-cosmos/tests/integration"
 	"github.com/bcp-innovations/hyperlane-cosmos/util"
 	"github.com/bcp-innovations/hyperlane-cosmos/x/core/01_interchain_security/types"
-	sdk "github.com/cosmos/cosmos-sdk/types"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 )
@@ -31,6 +31,12 @@ TEST CASES - message_id_multisig.go
 */
 
 var _ = Describe("message_id_multisig.go", Ordered, func() {
+	var s *i.KeeperTestSuite
+
+	BeforeEach(func() {
+		s = i.NewCleanChain()
+	})
+
 	It("Validate (invalid) invalid validators", func() {
 		// Arrange
 		validators := []string{
@@ -155,7 +161,7 @@ var _ = Describe("message_id_multisig.go", Ordered, func() {
 		metadata := bytesFromHexString("")
 
 		// Act
-		verify, err := messageIdMultisigIsm.Verify(sdk.Context{}, metadata, util.HyperlaneMessage{})
+		verify, err := messageIdMultisigIsm.Verify(s.Ctx(), metadata, util.HyperlaneMessage{})
 
 		// Assert
 		Expect(err.Error()).To(Equal("invalid metadata length: got 0, expected at least 68 bytes"))
@@ -197,7 +203,7 @@ var _ = Describe("message_id_multisig.go", Ordered, func() {
 		metadata.Signatures = signatures
 
 		// Act
-		verify, err := merkleRootMultisigIsm.Verify(sdk.Context{}, metadata.Bytes(), message)
+		verify, err := merkleRootMultisigIsm.Verify(s.Ctx(), metadata.Bytes(), message)
 
 		// Assert
 		Expect(err.Error()).To(Equal("invalid signatures length in metadata"))
@@ -237,7 +243,7 @@ var _ = Describe("message_id_multisig.go", Ordered, func() {
 		metadata.Signatures = signatures
 
 		// Act
-		verify, err := messageIdMultisigIsm.Verify(sdk.Context{}, metadata.Bytes(), message)
+		verify, err := messageIdMultisigIsm.Verify(s.Ctx(), metadata.Bytes(), message)
 
 		// Assert
 		Expect(err.Error()).To(Equal("threshold can not be reached"))
@@ -275,7 +281,7 @@ var _ = Describe("message_id_multisig.go", Ordered, func() {
 		metadata.Signatures = signatures
 
 		// Act
-		verify, err := messageIdMultisigIsm.Verify(sdk.Context{}, metadata.Bytes(), message)
+		verify, err := messageIdMultisigIsm.Verify(s.Ctx(), metadata.Bytes(), message)
 
 		// Assert
 		Expect(err.Error()).To(Equal("failed to recover validator signature: invalid signature recovery id"))
@@ -315,7 +321,7 @@ var _ = Describe("message_id_multisig.go", Ordered, func() {
 		metadata.Signatures = signatures
 
 		// Act
-		verify, err := messageIdMultisigIsm.Verify(sdk.Context{}, metadata.Bytes(), message)
+		verify, err := messageIdMultisigIsm.Verify(s.Ctx(), metadata.Bytes(), message)
 
 		// Assert
 		Expect(err).To(BeNil())
@@ -362,7 +368,7 @@ var _ = Describe("message_id_multisig.go", Ordered, func() {
 		metadata.Signatures = duplicatedSignatures
 
 		// Act
-		verify, err := messageIdMultisigIsm.Verify(sdk.Context{}, metadata.Bytes(), message)
+		verify, err := messageIdMultisigIsm.Verify(s.Ctx(), metadata.Bytes(), message)
 
 		// Assert
 		Expect(err).To(BeNil())
@@ -409,9 +415,10 @@ var _ = Describe("message_id_multisig.go", Ordered, func() {
 		metadata.Signatures = signatures
 
 		// Act
-		verify, err := messageIdMultisigIsm.Verify(sdk.Context{}, metadata.Bytes(), message)
+		verify, err := messageIdMultisigIsm.Verify(s.Ctx(), metadata.Bytes(), message)
 
 		// Assert
+		Expect(int(s.Ctx().GasMeter().GasConsumed())).To(Equal(3000))
 		Expect(err).To(BeNil())
 		Expect(verify).To(BeTrue())
 	})